literate: move class
