+++ base/src/main/java/org/aya/concrete/remark/LiterateConsumer.java
@@ -1,10 +1,9 @@
 // Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli.literate;
+package org.aya.concrete.remark;
 
 import kala.collection.immutable.ImmutableSeq;
 import kala.collection.mutable.MutableList;
-import org.aya.concrete.remark.Literate;
 import org.jetbrains.annotations.MustBeInvokedByOverriders;
 import org.jetbrains.annotations.NotNull;
 
@@ -51,13 +50,4 @@ public interface LiterateConsumer extends Consumer<Literate> {
     }
   }
 
-  record Highlights(@NotNull ImmutableSeq<HighlightInfo> highlights) implements LiterateConsumer {
-    @Override public void accept(@NotNull Literate literate) {
-      if (literate instanceof Literate.CodeBlock codeBlock && codeBlock.isAya() && codeBlock.sourcePos != null) {
-        var hl = highlights.filter(x -> codeBlock.sourcePos.containsIndex(x.sourcePos()));
-        codeBlock.highlighted = FaithfulDistiller.highlight(codeBlock.raw, codeBlock.sourcePos.tokenStartIndex(), hl);
-      }
-      LiterateConsumer.super.accept(literate);
-    }
-  }
 }
